requirements.txt
asn1crypto==0.24.0
beautifulsoup4==4.6.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cryptography==2.4.2
Django==2.1.3
django-cors-headers==2.4.0
django-modelcluster==4.3
django-taggit==0.23.0
django-treebeard==4.3
djangorestframework==3.9.0
draftjs-exporter==2.1.5
gunicorn==19.9.0
html5lib==1.0.1
idna==2.7
Pillow==5.3.0
pycparser==2.19
PyMySQL==0.9.2
pytz==2018.7
requests==2.20.1
six==1.11.0
Unidecode==1.0.23
urllib3==1.24.1
wagtail==2.3
webencodings==0.5.1
Willow==1.1
